finish moving rest of opcodes
